Update: Modularise PrettyDiff component
