+++ src/components/alexandria/PrettyDiff/styles.scss
@@ -1,6 +1,6 @@
-@import '../border';
-@import '../color';
-@import '../font-size';
+@import 'styles/border';
+@import 'styles/color';
+@import 'styles/font-size';
 
 .pretty-diff-component {
   border: $border-lighter;
