+++ src/components/core.js
@@ -48,7 +48,7 @@ import Grid from 'components/alexandria/Grid/component';
 import GridCell from 'components/alexandria/GridCell/component';
 import GridRow from 'components/alexandria/GridRow/component';
 import PageTitle from 'components/alexandria/PageTitle/component';
-import PrettyDiff from 'components/alexandria/PrettyDiffComponent';
+import PrettyDiff from 'components/alexandria/PrettyDiff/component';
 import Search from 'components/alexandria/SearchComponent';
 import Slicey from 'components/alexandria/SliceyComponent';
 import Spinner from 'components/alexandria/SpinnerComponent';
